+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Peripheral proxy now directional
+
 ## [1.4.0] - 2023-09-14
 
 ### BREAKING CHANGE